+++ app/pb_kits/playbook/pb_progress_simple/docs/_progress_simple_align.jsx
@@ -1,0 +1,28 @@
+import React from 'react'
+import { ProgressSimple } from '../..'
+
+const ProgressSimpleAlign = () => {
+  return (
+    <div>
+      <ProgressSimple
+          align="left"
+          percent={45}
+          width="100px"
+      />
+      <br />
+      <ProgressSimple
+          align="center"
+          percent={45}
+          width="100px"
+      />
+      <br />
+      <ProgressSimple
+          align="right"
+          percent={45}
+          width="100px"
+      />
+    </div>
+  )
+}
+
+export default ProgressSimpleAlign