+++ app/pb_kits/playbook/pb_progress_simple/docs/_progress_simple_muted.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import { ProgressSimple } from '../..'
+
+const ProgressSimpleMuted = () => {
+  return (
+    <div>
+      <ProgressSimple
+          muted
+          percent={68}
+      />
+    </div>
+  )
+}
+
+export default ProgressSimpleMuted